finish basic restAPI

+++ src/controllers/employee.controller.ts
@@ -7,23 +7,27 @@ import {
   Where,
 } from '@loopback/repository';
 import {
-  post,
-  param,
+  del,
   get,
   getModelSchemaRef,
+  param,
   patch,
+  post,
   put,
-  del,
   requestBody,
   response,
 } from '@loopback/rest';
-import {Employee} from '../models';
+import {Employee, EmployeeRequest, User} from '../models';
 import {EmployeeRepository} from '../repositories';
-
+import {UserRepository} from './../repositories/user.repository';
 export class EmployeeController {
+  bcrypt = require('bcrypt');
+  saltRounds = 10;
   constructor(
     @repository(EmployeeRepository)
-    public employeeRepository : EmployeeRepository,
+    public employeeRepository: EmployeeRepository,
+    @repository(UserRepository)
+    public userRepository: UserRepository,
   ) {}
 
   @post('/employees')
@@ -35,16 +39,35 @@ export class EmployeeController {
     @requestBody({
       content: {
         'application/json': {
-          schema: getModelSchemaRef(Employee, {
+          schema: getModelSchemaRef(EmployeeRequest, {
             title: 'NewEmployee',
             exclude: ['id'],
           }),
         },
       },
     })
-    employee: Omit<Employee, 'id'>,
-  ): Promise<Employee> {
-    return this.employeeRepository.create(employee);
+    req: Omit<EmployeeRequest, 'id'>,
+  ): Promise<any> {
+    //  console.log(req);
+    const {user, ...employee} = req;
+
+    const newUser = await this.userRepository.createFromEmployee(user);
+
+    if (newUser) {
+      if (newUser.id != undefined) {
+        employee.userId = newUser.id;
+
+        try {
+          const addEmp = await this.employeeRepository.create(employee);
+          return addEmp;
+        } catch (e) {
+          const deleteUser = this.userRepository.delete(newUser);
+          return 'add new employee failed';
+        }
+      }
+    } else {
+      return `User existed`;
+    }
   }
 
   @get('/employees/count')
@@ -52,9 +75,7 @@ export class EmployeeController {
     description: 'Employee model count',
     content: {'application/json': {schema: CountSchema}},
   })
-  async count(
-    @param.where(Employee) where?: Where<Employee>,
-  ): Promise<Count> {
+  async count(@param.where(Employee) where?: Where<Employee>): Promise<Count> {
     return this.employeeRepository.count(where);
   }
 
@@ -76,24 +97,24 @@ export class EmployeeController {
     return this.employeeRepository.find(filter);
   }
 
-  @patch('/employees')
-  @response(200, {
-    description: 'Employee PATCH success count',
-    content: {'application/json': {schema: CountSchema}},
-  })
-  async updateAll(
-    @requestBody({
-      content: {
-        'application/json': {
-          schema: getModelSchemaRef(Employee, {partial: true}),
-        },
-      },
-    })
-    employee: Employee,
-    @param.where(Employee) where?: Where<Employee>,
-  ): Promise<Count> {
-    return this.employeeRepository.updateAll(employee, where);
-  }
+  // @patch('/employees')
+  // @response(200, {
+  //   description: 'Employee PATCH success count',
+  //   content: {'application/json': {schema: CountSchema}},
+  // })
+  // async updateAll(
+  //   @requestBody({
+  //     content: {
+  //       'application/json': {
+  //         schema: getModelSchemaRef(Employee, {partial: true}),
+  //       },
+  //     },
+  //   })
+  //   employee: Employee,
+  //   @param.where(Employee) where?: Where<Employee>,
+  // ): Promise<Count> {
+  //   return this.employeeRepository.updateAll(employee, where);
+  // }
 
   @get('/employees/{id}')
   @response(200, {
@@ -106,7 +127,8 @@ export class EmployeeController {
   })
   async findById(
     @param.path.number('id') id: number,
-    @param.filter(Employee, {exclude: 'where'}) filter?: FilterExcludingWhere<Employee>
+    @param.filter(Employee, {exclude: 'where'})
+    filter?: FilterExcludingWhere<Employee>,
   ): Promise<Employee> {
     return this.employeeRepository.findById(id, filter);
   }
@@ -144,7 +166,29 @@ export class EmployeeController {
   @response(204, {
     description: 'Employee DELETE success',
   })
-  async deleteById(@param.path.number('id') id: number): Promise<void> {
-    await this.employeeRepository.deleteById(id);
+  async deleteById(
+    @param.path.number('id') id: number,
+  ): Promise<string | void> {
+    console.log(`[deleteById] req: ${id}`);
+    try {
+      console.log(`[findById Employee] req: ${id}`);
+      let employee: Employee = await this.employeeRepository.findById(id);
+
+      if (employee) {
+        console.log(`[findById User] req: ${employee.userId}`);
+        let user: User = await this.userRepository.findById(employee.userId);
+
+        if (user) {
+          console.log(`[deleteById Employee] req: ${id}`);
+          await this.employeeRepository.deleteById(id);
+          console.log(`[findById User] req: ${user.id}`);
+          await this.userRepository.deleteById(user.id);
+        } else {
+          return `User with id ${employee.id} doesn't exist`;
+        }
+      }
+    } catch (e) {
+      return `Failed to find employee with id ${id} !`;
+    }
   }
 }
